Add caches command
* Add caches command to cli
* Add caches method to CapistranoDeployer
* Add a "split token", which we use to artificially introduce strings to
  cap tasks that we use to split output later.

We get the list by simply using cap to perform an ls on the releases
directory at each machine. The capistrano deployer then parses this
output into an array of strings and returns that to the cli.

At this stage, we don't do anything special with the directory listing,
so it is only timestamps of when cap performed the deploy, e.g.
20171019013839. At a later time when there is a local history to lean
on, this output can be made more useful.

Finally, we investigating dropping into the cap environment from
directly within ruby, which would let us avoid the split_token nonsense.
However, this is not a documented feature of cap/rake, and the path
to hacking this functionality together was non-trivial and appeared
unstable.