Fix bug in GitRunner#safe_checkout dir name
Previously, #safe_checkout would use git's default behavior of naming
the created folder by the name of the git repository. This caused a bug
because we did not discover that name.

This changes that behavior to always name the created folder 'fauxpaas'.
There is no danger of collision because this is done in a tmpdir.

Additionally:

* Parameterizes GitRunner and thus subclasses
* GitRunner uses an injected filesystem instead of calling Dir.chdir
  directly.