Change expected paths
This commit seeks to finalize the paths that we expect to use in
production environments. It creates a /data/ folder that holds the
instance data previously under /deploy. The deploy directory now
explicitly contains code-like files only.

Stages were of particular concern; capistrano does not allow nesting of
stages, nor of providing their paths on the command line. It only allows
moving the "stages" directory. To prevent this from being one more place
that may require changes, we use symlinks.

Releases are now under a /releases/ folder, similar to a log. Thus,
creating new releases will no longer modify the SHA of an instance.

We also move the infrastructure and deploy configurations for the test
repositories out of this repo. While this is a logical home for them,
I felt that their inclusion here was confusing at this time.

Finally, we ignore non-test files in the data and releases directories.