Overhaul configuration to ettin+canister
Previously, our configuration was all over the place, and
the model was quickly becoming unsustainable. This effort
attempts to get all of that configuration into one manageable
location that can be easily changed as needed.

This commit also fixes the following classes to properly
expect their dependencies to be injected:

* ArchiveReference
* CapRunner
* DeployConfig
* FileInstanceRepo
* GitRunner
* ReleaseBuilder

Lastly, we were forced to remove the rake task for spec:integration
and spec:all because rspec no longer feels compelled to allow the
subcontexts of the integration_spec.rb to run one after another
unless you specify the file exactly. I don't know why.