Create integration tests for (no)rails deploys
* Removes old test-rails and test-norails
* Adds half-instance artifact test-persistence for unit tests of
  FileInstanceRepo.
* Creates integration tests for a rails deploy
* Creates integration tests for a non-rails deploy
* Supports ERB in instance.yml
* Removes troublesome gitignore files
* Adds uncommitted stage files
* Add rake tasks for running unit or integration tests

In order to enable this functionality, we had to add a number of other
branches to the project. These branches are required for the integration
tests to run. They are:

* test/norails/source
* test/norails/deploy
* test/norails/shared
* test/norails/unshared
* test/rails/source
* test/rails/deploy
* test/rails/shared
* test/rails/unshared

The integration tests work by considering these branches as their
sources.  One should perform a git pull of each of these, else the
integration tests will fail.  Additionally, there are now rake tasks
for running the unit test suite vs the integration test suite.

The integration tests themselves work by making a single deploy of the
test case (e.g. test-rails or test-norails), and then making a series
of assertions on the deployed instance. Thus, if the deploy fails
entirely, all fo the tests will likely fail. This was a compromise to
keep the test runtime from ballooning.