Decouple CLI from the core logic
The goal of this set of changes was to completely separate the CLI from
the rest of the application. This was primarily motivated by the
increasing frequency of broken integration tests masquerading as unit
tests. Several areas were identified as contributing heavily to the
coupling between the CLI and the core of the application:

  * Construction of the authorization policy object
  * Complexity in finding, printing, and redirecting output
  * Testing of Command behavior lived in the CLI's tests

== Logging ==
Previously, I used puts for all of the program's output. This caused
the tests to print useless text. To begin addressing this, I moved
everything over to a centrally-defined ruby logger instance.

This change is largely positive, but it does make the caches, releases,
and default branch output overly verbose. I are also not capturing
system or git output any more cleanly than before.

This allowed the removal of the VerboseRunner.

== Invoker ==
The invoker is a standard part of the command pattern which I
previously omitted. Moving command invocation out of the cli/client
into the invoker enables simplicity.

== Misc CLI Cleanup ==
I discovered that the initialize method can be used to run code before
every command. Setup was moved here.

Deploy's "reference" parameter was adjusted to function like the
parameters of other commands.

== Policy and Authorization ==
Several changes were done to fauxpaas's authorization system code.

The repo -> factory -> policy design has been modified. The "factory"
piece is now the AuthService, which uses Policy internally to decide
whether or not to authorize actions. The actual decision-making is still
wholly within Policy, but Policy is otherwise hidden from callers.

It's now the commands that pass the action, entity, and user to the
AuthService rather than the CLI.

FilePolicyFactoryRepo has been renamed to FilePermissionsRepo, and now
simply produces the data the AuthService needs rather than minting the
AuthService itself.

Actions have been reified to symbols, instead of having a unique method
on Policy for each command.

The Command#run method has been removed. The responsibility for only
running valid, authorized commands falls to the Invoker.