Fix error on missing assets:precompile task
Resolves AEIM-1270

The task set to run after assets:precompile (defined in
deploy/cap/assets.rb) would cause an _immediate_ error as the
assets:precompile task was not defined.

By adding an empty task, this error condition is removed. The dummy task
does not override the actual task regardless of definition order.