AEIM-1519 - Remove platform dependency of PassthroughRunner
The tests for PassthroughRunner check for platform-specific output of
the ls command on a nonexistent file as a way to get output to stderr.
Here, we change to use standard redirection and fixed output with echo.

Because the implementation always merges stdout and stderr, this also
changes the examples for separate streams to exclude output to the
other stream and marks those examples as pending. They could be enabled
if/when separate streams were available (maybe via an option).