AEIM-1359: Artifact (WIP)
This is a start at refactoring the data clump of source_path /
shared_path / unshared_path so that we can have a way to vary what we're
doing for AEIM-1361, etc (including various things in the artifact /
built release).

The next step here is probably to build the Artifact just using the
Signature and then have the Artifact itself (for now) handle the
checkout - AEIM-1357 will abstract that part.

Then we can probably collapse Release and ReleaseBuilder into its own
class.

Once that's done and we implement AEIM-1361/AEIM-1362, the Artifact can
just need to give Cap a single path where the artifact can be found
instead of three separate paths.