AEIM-1365 - bundle workarounds; integration test for assets
- Several workarounds to get bundler working with specs in a
non-breaking and non-slow way. This would be inappropriate for real
usage, but a working bundle install without capistrano is both a
prerequisite and out of scope for this issue. In particular, it just
uses 'bundle install' for now so it uses the system gem cache, which is
much faster.

I removed capistrano-bundler, since it should no longer be needed (the
identical change is also needed for AEIM-1363).

I also removed the version restriction on rake in fauxpaas
so it can use the current rake.

- Fixed the after_build command so that db:migrate works as expected.

Remaining issues:

- artifact_builder specs are very slow (presumably because it's running the
external stuff)

- non-rails apps still run rake asset:precompile, but the specs pass
anyway because ArtifactBuilder doesn't check if it the command actually
works.