Fix bug caused by unused bundle package
The 'bundle package' command that was run caused bundler to request
sudo packages in linux, but not mac. This commit fixes that issue
by simply removing the command and its corresponding test.

As the following 'bundle install' command did not include '--local',
the cached gems were not used. As such, there should be no adverse
effects.