Add TaskFile::TaskSpec
A subclass (or struct as the case may be) is preferable here as it
avoids passing around hashes. We add the factory methods to the task
classes to complete the refactor.

Adds TaskFile::from_path, which contains the previous behavior of ::new.
::new now expects an array of hashes (raw tasks) as returned by the yaml
parser. This avoids the need for FakeFS in the specs for this unit.