Fix incorrect setting of current symlink
Task::SetCurrent was relying on $PWD, and this was being resolved on
the local host rather than the remote one, causing the symlink to always
be broken.

We didn't discover this behavior in the systemd restart action, because
systemd doesn't error until after the restart command has finished.

This commit changes the logic to always get absolute paths from the
Release instance.