AEIM-1847: remove extraneous command.first_run? in Moku::CLI init
This caused an exception here as command was not bound the expected way.
It might work to check using this.first_run? instead, but given that the
init command has a precheck that is checking first_run? this seemed
redundant.

There are not currently any tests for Moku::CLI as it's not obvious how
to isolate it for unit testing. Tests for Command::Init are also missing
but wouldn't have caught this particular bug; those would probably be
worth adding when we encounter a bug in Command::Init itself. As there
starts being more complexity in Moku::CLI we should think more about how
to test it.