AEIM-1847: Add Sites.to_h method, use it in Moku::Pipeline::Init
This gives us a single, testable implementation for converting between
the various representations of node to datacenter hashes.

n.b. I added the "Symbol" case in Sites because some of the node names
from the puppet json were coming in as Symbols; it seems easier to just
stringify them than to chase down in what cases that might happen.