Reverse pipeline -> command coupling
The components of the command pattern are the command, the receiver, the
invoker, and the client.  We've basically split the _command_ in two.
Command::Deploy + Pipeline::Deploy constitute a single,
concrete _command_ in the traditional pattern, with the _receiver_ being
the logical instance rather than anything reified in moku.

This change embraces this coupling between Command and Pipeline
subclasses. By having the command instantiate the pipeline with all of
the information that it needs, we better match the traditional command
pattern.  Removing the factory from Pipeline also reduces indirection
that was not providing us anything.