AEIM-2033 - Work around nested single quotes
Commands to be issued under bash over ssh are single quoted. The
original ruby command to make the symlinks was single quoted and
conflicted with this. This doesn't solve the larger challenge, but
changes the outer ruby -e single quotes to doubles and uses alternate
quotes in the expression to avoid needing any singles.