AEIM-2507 - Pass instance name, uid, gid to Docker tasks
This is a bit of a workaround for now -- the tasks did not have access
to instance information, and it's not totally clear where it should
reside. In the absence of a cleaner reconfiguration of the
responsibilities, we add plan factories that capture the Instance and
pass it down to the Docker-specific plans, and then they pass it down to
the tasks. The deployment config has the uid/gid, so it is exposed on
the Instance directly. Much of this stuff could be moved around...

Additionally:

 - The Docker command exit status and logging are cleaned up
 - The Plan and Task pieces are disambiguated